v1.2.5

+++ scripts/PyPi_push.ps1
@@ -1,4 +1,6 @@
-Remove-Item "dist" -Recurse
-python setup.py bdist_wheel
+if (Test-Path "dist") {
+  Remove-Item "dist" -Recurse
+}
+python -m build --wheel
 twine check dist/*
 twine upload dist/*
